move softmaxcrossetropyloss to math folder
