+++ docs/Changelog.md
@@ -14554,7 +14554,3 @@ This version of the operator has been available since version 12 of the default 
 <dd>Constrain input and output types to float tensors.</dd>
 </dl>
 
-#### Function
-
-The Function can be represented as a function.
-
