+++ docs/Operators.md
@@ -12129,52 +12129,6 @@ This version of the operator has been available since version 11 of the default 
 </dl>
 
 
-#### Examples
-
-<details>
-<summary>range_float_type_positive_delta</summary>
-
-```python
-node = onnx.helper.make_node(
-    'Range',
-    inputs=['start', 'limit', 'delta'],
-    outputs=['output'],
-)
-
-start = np.float32(1)
-limit = np.float32(5)
-delta = np.float32(2)
-
-output = np.arange(start, limit, delta, dtype=np.float32)  # expected output [1.0, 3.0]
-expect(node, inputs=[start, limit, delta], outputs=[output],
-       name='test_range_float_type_positive_delta')
-```
-
-</details>
-
-
-<details>
-<summary>range_int32_type_negative_delta</summary>
-
-```python
-node = onnx.helper.make_node(
-    'Range',
-    inputs=['start', 'limit', 'delta'],
-    outputs=['output'],
-)
-
-start = np.int32(10)
-limit = np.int32(6)
-delta = np.int32(-3)
-
-output = np.arange(start, limit, delta, dtype=np.int32)  # expected output [10, 7]
-expect(node, inputs=[start, limit, delta], outputs=[output],
-       name='test_range_int32_type_negative_delta')
-```
-
-</details>
-
-
 ### <a name="Reciprocal"></a><a name="reciprocal">**Reciprocal**</a>
 
   Reciprocal takes one input data (Tensor<T>) and produces one output data
@@ -17545,10 +17499,6 @@ This version of the operator has been available since version 12 of the default 
 <dt><tt>T</tt> : tensor(float16), tensor(float), tensor(double)</dt>
 <dd>Constrain input and output types to float tensors.</dd>
 </dl>
-
-#### Function
-
-The Function can be represented as a function.
 
 
 #### Examples
